data_challenge to rtmodel - raising errors

+++ jasmine/files_organizer/data_challenge_to_RTModel.py
@@ -11,16 +11,16 @@ def creating_directories(*, data_challenge_lc_number_,
     """
     event_folder_ = f'{folder_path_}/event_{data_challenge_lc_number_:03}'
     data_folder_ = f'{folder_path_}/event_{data_challenge_lc_number_:03}/Data'
-    try:
+    if os.path.isdir(event_folder_):
+        raise FileExistsError(f'Directory {event_folder_} already exists!')
+    else:
         os.mkdir(event_folder_)
         print(f'Folder {event_folder_} created!')
-    except FileExistsError:
-        print(f'Directory {event_folder_} already exists!')
-    try:
+    if os.path.isdir(data_folder_):
+        raise FileExistsError(f'Directory {data_folder_} already exists!')
+    else:
         os.mkdir(data_folder_)
         print(f'Folder {data_folder_} created!')
-    except FileExistsError:
-        print(f'Directory {data_folder_} already exists!')
 
 
 def creating_files(*, data_challenge_lc_number_,
@@ -36,11 +36,14 @@ def creating_files(*, data_challenge_lc_number_,
     for filter_ in filters:
         lightcurve_data_df = the_lightcurve.lightcurve_data(filter_=filter_, folder_path_=data_input_folder_path_)
         lightcurve_data_df_out = lightcurve_data_df[['Magnitude', 'Error', 'days']]
-        with open(f'{data_folder_}/Roman{filter_}.dat', 'w') as f:
-            f.write('# Mag err HJD-2450000\n')
-        lightcurve_data_df_out.to_csv(f'{data_folder_}/Roman{filter_}.dat', sep=' ',
-                                      index=False, header=False, mode='a')
-        print(f'File Roman{filter_}.dat created in {data_folder_}.')
+        if os.path.exists(f'{data_folder_}/Roman{filter_}.dat'):
+            raise FileExistsError(f'File Roman{filter_}.dat already exists in directory {data_folder_}')
+        else:
+            with open(f'{data_folder_}/Roman{filter_}.dat', 'w') as f:
+                f.write('# Mag err HJD-2450000\n')
+            lightcurve_data_df_out.to_csv(f'{data_folder_}/Roman{filter_}.dat', sep=' ',
+                                          index=False, header=False, mode='a')
+            print(f'File Roman{filter_}.dat created in {data_folder_}.')
 
     # Create the file with the event information
     event_ra = the_lightcurve.event_info.event_right_ascension__deg
@@ -48,9 +51,12 @@ def creating_files(*, data_challenge_lc_number_,
     event_coordinates = SkyCoord(event_ra, event_dec, unit=(units.deg, units.deg), obstime='J2000')
     rtmodel_event_coordinates = (event_coordinates.to_string('hmsdms').replace('h', ':').replace('m', ':')
                                  .replace('d', ':').replace('s', ''))
-    with open(f'{data_folder_}/event.coordinates', 'w') as f:
-        f.write(f'{rtmodel_event_coordinates}')
-    print(f'File event.coordinates created in {data_folder_}.')
+    if os.path.exists(f'{data_folder_}/event.coordinates'):
+        raise FileExistsError(f'File event.coordinates already exists in directory {data_folder_}')
+    else:
+        with open(f'{data_folder_}/event.coordinates', 'w') as f:
+            f.write(f'{rtmodel_event_coordinates}')
+        print(f'File event.coordinates created in {data_folder_}.')
 
 
 if __name__ == '__main__':
@@ -63,6 +69,3 @@ if __name__ == '__main__':
                    output_folder_path_=folder_to_be_created,
                    master_input_folder_path_=input_folder,
                    data_input_folder_path_=input_folder)
-
-
-
